Upd Render System

+++ src/Components/TransformComponent.h
@@ -7,7 +7,7 @@ struct TransformComponent : public Component<TransformComponent>
 {
 	glm::vec2 position{ 0, 0 };
 	glm::vec2 scale{ 1, 1 };
-	float rotation{ 0.0f };
+	double rotation{ 0.0f };
 
 	TransformComponent() = default;
